added interface impls

+++ src/main/java/com/freeuni/quizapp/model/QuizResult.java
@@ -8,12 +8,12 @@ public class QuizResult {
     private int quizId;
     private int score;
     private int totalQuestions;
-    private long timeTakenSeconds;
+    private int timeTakenSeconds;
     private boolean isPracticeMode;
     private Timestamp completedAt;
 
     public QuizResult(int id, int userId, int quizId, int score,
-                      int totalQuestions, long timeTakenSeconds, boolean isPracticeMode,
+                      int totalQuestions, int timeTakenSeconds, boolean isPracticeMode,
                       Timestamp completedAt) {
         this.id = id;
         this.userId = userId;
@@ -70,11 +70,11 @@ public class QuizResult {
         this.totalQuestions = totalQuestions;
     }
 
-    public long getTimeTakenSeconds() {
+    public int getTimeTakenSeconds() {
         return timeTakenSeconds;
     }
 
-    public void setTimeTakenSeconds(long timeTakenSeconds) {
+    public void setTimeTakenSeconds(int timeTakenSeconds) {
         this.timeTakenSeconds = timeTakenSeconds;
     }
